Q now has Read([]byte) Write([]byte) interface
Previously we were retrieving a byte slice from the q, writing to it and
then confirming we had finished writing. We would read from the queue in
the same way. On my chromebook this implementation managed 100MM
write/reads in ~18 seconds.

The new interface allows the client code to cache (or not) it's own
buffer []byte and pass that into the queue for writing. On my chromebook
this implementation manages 100MM write/reads in ~13 seconds.