atomicint.LazyAdd() -> fatomic.LazyStore()
This removes a single method call and seemed to improve performance on
my Chromebook. Needs to be double checked on a real computer. This is a
manual inline that should really be done by the compiler. I have the
feelling we are getting into marginal improvement hacks where the
readability/performance tradeoff is going sour. Might be time to move
on.